requirements.txt
appdirs==1.4.3
Django==1.11
django-filter==1.0.4
django-mssql==1.8
django-pymssql==1.7.1
django-python3-ldap==0.9.14
django-suit==0.2.25
djangorestframework==3.6.3
ldap3==2.2.0
Markdown==2.6.8
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pyldap==2.4.35.1
pymssql==2.1.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
